fix: added the correct image and text when there are no classes to display and refactored the code to conform with the norm across the rest of the project
