+++ packages/uni_app/lib/view/academic_path/widgets/no_classes_widget.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'package:uni/generated/l10n.dart';
+import 'package:uni/view/common_widgets/expanded_image_label.dart';
+
+class NoClassesWidget extends StatelessWidget {
+  const NoClassesWidget({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return ImageLabel(
+      imagePath: 'assets/images/school.png',
+      label: S.of(context).no_class,
+      labelTextStyle: TextStyle(
+        fontWeight: FontWeight.bold,
+        fontSize: 18,
+        color: Theme.of(context).colorScheme.primary,
+      ),
+      sublabel: S.of(context).no_classes_this_week,
+      sublabelTextStyle: Theme.of(context).textTheme.bodyLarge,
+    );
+  }
+}
